refactor: layout-setting
